SRC/requirements.txt
Flask==1.1.1
SQLAlchemy==1.3.13
Keras==2.3.1
numpy==1.17.4
python-dotenv==0.11.0
requests==2.21.0
pandas==1.0.1
jsonschema==3.2.0
PyMySQL==0.9.3
textblob==0.15.3
gunicorn==20.0.4
tensorflow==1.14.0